# Conflicts:
#	wrappersnap/launcher/src/main/java/wasdi/dataproviders/LpDaacProviderAdapter.java
#	wrappersnap/wasdishared/src/main/java/wasdi/shared/queryexecutors/QueryTranslator.java
#	wrappersnap/wasdishared/src/main/java/wasdi/shared/queryexecutors/lpdaac/QueryExecutorLpDaac.java
#	wrappersnap/wasdishared/src/main/java/wasdi/shared/queryexecutors/terrascope/QueryExecutorTerrascope.java

+++ wrappersnap/launcher/src/main/java/wasdi/dataproviders/CDS2ProviderAdapter.java
@@ -23,13 +23,34 @@ public class CDS2ProviderAdapter extends PythonBasedProviderAdapter {
 
 	public CDS2ProviderAdapter() {
 		super();
-		m_sDataProviderCode = "CDS";
 	}
 	
 	@Override
 	public long getDownloadFileSize(String sFileURL) throws Exception {
 		WasdiLog.debugLog("CDS2ProviderAdapter.getDownloadFileSize: download file size not available in CDS");
 		return 0;
+	}
+	
+	@Override
+	protected Map<String, Object> fromWasdiPayloadToObjectMap(String sUrl) {
+		String sDecodedUrl = decodeUrl(sUrl);
+
+		String sPayload = null;
+		if (!Utils.isNullOrEmpty(sDecodedUrl)) {
+			String[] asTokens = sDecodedUrl.split("payload=");
+			if (asTokens.length == 2) {
+				sPayload = asTokens[1];
+				WasdiLog.debugLog("CDS2ProviderAdapter.fromtWasdiPayloadToObjectMap json string: " + sPayload);
+				return JsonUtils.jsonToMapOfObjects(sPayload);
+			}
+			WasdiLog.debugLog("CDS2ProviderAdapter.fromtWasdiPayloadToObjectMap. Payload not found in url " + sUrl);
+		}
+		
+		HashMap<String, Object> aoReturnMap = new HashMap<>();
+		aoReturnMap.put("url", sUrl);
+		
+		WasdiLog.warnLog("CDS2ProviderAdapter.fromtWasdiPayloadToObjectMap. Decoded url is null or empty " + sUrl);
+		return aoReturnMap;
 	}
 	
 	@Override
@@ -111,13 +132,20 @@ public class CDS2ProviderAdapter extends PythonBasedProviderAdapter {
 				WasdiLog.errorLog("CDS2ProviderAdapter.executeDownloadFile: path to the downloaded file is null or empty");
 			}
 			
-		} catch(Exception oEx) {
+		} 
+		catch(Exception oEx) {
 			WasdiLog.errorLog("CDS2ProviderAdapter.executeDonwloadFile: error ", oEx);
-		} finally {
-			if (!Utils.isNullOrEmpty(sInputFullPath))
-				FileUtils.deleteQuietly(new File(sInputFullPath));
-			if (!Utils.isNullOrEmpty(sOutputFullPath))
-				FileUtils.deleteQuietly(new File(sOutputFullPath));
+		} 
+		finally {
+			if (WasdiConfig.Current.dockers.removeParameterFilesForPythonsShellExec) {
+				if (!Utils.isNullOrEmpty(sInputFullPath))
+					FileUtils.deleteQuietly(new File(sInputFullPath));
+				if (!Utils.isNullOrEmpty(sOutputFullPath))
+					FileUtils.deleteQuietly(new File(sOutputFullPath));
+			}
+			else {
+				WasdiLog.infoLog("CDS2ProviderAdapter.executeDonwloadFile: removeParameterFilesForPythonsShellExec is  FALSE, we keep the parameter files");
+			}
 		}
 		
 		return sResultDownloadedFilePath;
